refactor according to comments in github
