--- PivotalTracker.properties
@@ -1,7 +1,0 @@
-browser=chrome
-
-baseUrl=https://www.pivotaltracker.com/
-
-implicitWait=30
-explicitWait=40
-waitSleepTime=500
