+++ build.gradle
@@ -23,4 +23,23 @@ dependencies {
     compile group: 'io.github.bonigarcia', name: 'webdrivermanager', version: '1.7.2'
     compile group: 'log4j', name: 'log4j', version: '1.2.17'
     compile group: 'com.google.code.gson', name: 'gson', version: '2.8.5'
+    compile group: 'org.json', name: 'json', version: '20180813'
+}
+
+// Task to define arguments and configurations to properties file
+tasks.withType(Test) {
+
+    // enable TestNG support (default is JUnit)
+    useTestNG()
+    testLogging.showStandardStreams = true
+
+    // set JVM arguments for the features.test JVM(s)
+    scanForTestClasses = false
+
+    //turn off Gradle's HTML report to avoid replacing the reports generated by TestNG library:
+    reports.html.enabled = true
+
+    ignoreFailures = false
+
+    systemProperty "browser", browser
 }
