+++ gradle.properties
@@ -1,0 +1,2 @@
+#type of browser chrome/firefox
+browser=chrome
