+++ pivotaltracker.properties
@@ -1,0 +1,10 @@
+implicitWait=30
+explicitWait=40
+waitSleepTime=500
+
+baseUrl=https://www.pivotaltracker.com/
+
+#api
+apiUrl=services/v5/projects"
+accessToken=2e399ae84895eef55d322156e42f5c41
+contentType=application/json
