+++ src/main/java/core/utils/ConfigFileReader.java
@@ -26,6 +26,7 @@ import java.util.Properties;
  * @version 1.0
  */
 public final class ConfigFileReader {
+
     private static final String PIVOTAL_TRACKER_PROPERTIES = "PivotalTracker.properties";
     private Map<String, String> properties;
 
@@ -78,5 +79,32 @@ public final class ConfigFileReader {
             Log.getInstance().getLog().error(e.getMessage());
         }
         return properties;
+    }
+
+    /**
+     * Returns the token from the properties file.
+     *
+     * @return the token as string.
+     */
+    public String getAccessToken() {
+        return getProperties().get("accessToken");
+    }
+
+    /**
+     * Returns the url to request to the API from the properties file.
+     *
+     * @return the url as string.
+     */
+    public String getUrlApi() {
+        return getProperties().get("urlBase").concat(getProperties().get("apiUrl"));
+    }
+
+    /**
+     * Returns the content type from the properties file.
+     *
+     * @return the content type as string.
+     */
+    public String getContentType() {
+        return getProperties().get("contentType=application/json");
     }
 }
